ok, proceed with the reset-to-default-if-stubbed strategy.
because puma-dev mixes flags and commandline arguments, we're abusing os.Args in our tests. we can't just nuke all the flags, because then the flags that are passed in for testing are broken (e.g. we get no coverage). so, we need to record which commandline args have been provide that actually override the cached flag values, and surgically reset them.